Clean up of solution
- Added solution folders for each service
- Separated the message contracts according to service.
- Updated NServiceBus to RC2
- Cleaned up the container include for endpoints that use Autofac.
