+++ src/Billing.Events/OrderBilled.cs
@@ -1,7 +1,7 @@
-﻿namespace EShop.Messages.Events
-{
+﻿namespace Billing.Events
+{    
     using NServiceBus;
-    public class OrderPlaced : IEvent
+    public class OrderBilled : IEvent
     {
         public string OrderId { get; set; }
         public int ProductId { get; set; }
