--- src/EShop.Messages/Commands/StoreOrder.cs
@@ -1,11 +1,0 @@
-﻿namespace EShop.Messages.Commands
-{
-    using System;
-    using NServiceBus;
-    public class StoreOrder : ICommand
-    {
-        public string OrderId { get; set; }
-        public int ProductId { get; set; }
-        public DateTime OrderPlacedOn { get; set; }
-    }
-}
