+++ src/LoadGenerator/LoadGenerator.csproj
@@ -10,12 +10,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="7.0.0-rc0001" />
+    <PackageReference Include="NServiceBus" Version="7.0.0-rc0002" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\EShop.Messages\EShop.Messages.csproj" />
     <ProjectReference Include="..\ITOps.Shared\ITOps.Shared.csproj" />
+    <ProjectReference Include="..\Sales.Internal\Sales.Internal.csproj" />
   </ItemGroup>
 
 </Project>
