+++ src/Marketing.Events/ViewModelComposition/ProductsLoaded.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Marketing.Events.ViewModelComposition
+{
+    public class ProductsLoaded
+    {
+        public IDictionary<dynamic, dynamic> OrdersDictionary { get; set; }
+    }
+}
