+++ src/Marketing.Internal/RecordConsumerBehavior.cs
@@ -1,6 +1,7 @@
-﻿namespace EShop.Messages.Commands
+﻿using NServiceBus;
+
+namespace Marketing.Internal
 {
-    using NServiceBus;
     public class RecordConsumerBehavior : ICommand
     {
         public int ProductId { get; set; }
