+++ src/Sales.Events/OrderAccepted.cs
@@ -1,7 +1,8 @@
-﻿namespace EShop.Messages.Events
+﻿using NServiceBus;
+
+namespace Sales.Events
 {
-    using NServiceBus;
-    public class OrderBilled : IEvent
+    public class OrderAccepted : IEvent
     {
         public string OrderId { get; set; }
         public int ProductId { get; set; }
