+++ src/Sales.Events/OrderCancelled.cs
@@ -1,6 +1,7 @@
-﻿namespace EShop.Messages.Events
+﻿using NServiceBus;
+
+namespace Sales.Events
 {
-    using NServiceBus;
     public class OrderCancelled : IEvent
     {
         public string OrderId { get; set; }
