+++ src/Sales.Events/Sales.Events.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+  
+  <ItemGroup>
+    <PackageReference Include="NServiceBus" Version="7.0.0-rc0002" />
+  </ItemGroup>
+
+</Project>
