+++ src/Sales.Internal/PlaceOrder.cs
@@ -1,0 +1,10 @@
+﻿using NServiceBus;
+
+namespace Sales.Internal
+{
+    public class PlaceOrder : ICommand
+    {
+        public string OrderId { get; set; }
+        public int ProductId { get; set; }
+    }
+}
