+++ src/Sales.ViewModelComposition/ProductListGetHandler.cs
@@ -1,9 +1,10 @@
-﻿namespace Sales.ViewModelComposition
+﻿using Marketing.Events.ViewModelComposition;
+
+namespace Sales.ViewModelComposition
 {
     using System.Linq;
     using System.Collections.Generic;
     using System.Net.Http;
-    using EShop.Messages.ViewModelCompositionEvents;
     using ITOps.ViewModelComposition;
     using ITOps.ViewModelComposition.Json;
     using Microsoft.AspNetCore.Http;
