+++ src/Shipping.Internal/ItemRestocked.cs
